requirements.txt
# JobHunter AI - Dependencies

# Web scraping
playwright==1.49.0
beautifulsoup4==4.12.3
httpx==0.27.2
lxml==5.3.0

# AI
anthropic==0.39.0

# Notion integration
notion-client==2.2.1

# Google Drive
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Document generation
python-docx==1.1.2
weasyprint==62.3

# Data handling
pydantic==2.10.2
python-dateutil==2.9.0

# Environment and config
python-dotenv==1.0.1
pydantic-settings==2.6.1

# Utilities
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
